storing the new structure in this commit

+++ app/crud.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException
 from sqlalchemy.orm import Session
 
 from app import auth_utils, models, schemas
@@ -31,6 +32,43 @@ def get_user_by_email(email: str, db: Session) -> schemas.User:
     return db.query(models.User).filter(models.User.email == email).first()
 
 
+# User UPDATE
+def update_user_add_book(user_id: int, book_id: int, db: Session) -> schemas.User:
+    db_user = db.query(models.User).filter(models.User.id == user_id).first()
+    db_book = db.query(models.Book).filter(models.Book.id == book_id).first()
+    db_user.books.append(db_book)
+    db.commit()
+    db.refresh(db_user)
+    return db_user
+
+
+# User UPDATE
+def update_user_remove_boook(user_id: int, book_id: int, db: Session) -> schemas.User:
+    db_user = db.query(models.User).filter(models.User.id == user_id).first()
+    db_book = db.query(models.Book).filter(models.Book.id == book_id).first()
+    db_user.books.remove(db_book)
+    db.commit()
+    db.refresh(db_user)
+    return db_user
+
+
+# Book CREATE
+def create_book(used_id: int, book: schemas.BookCreate, db: Session) -> schemas.Book:
+    db_book = models.Book(**book.dict())
+    db_user = db.query(models.User).filter(models.User.id == used_id).first()
+    db.add(db_book)
+    db_book.users.append(db_user)
+    db.commit()
+    db.refresh(db_book)
+    return db_book
+
+
+# Book READ
+def read_book(user_id: int, book_id: int, db: Session) -> schemas.Book:
+    db_book = db.query(models.Book).filter(models.Book.id == book_id).first()
+    return db_book
+
+
 # Club CREATE
 def create_club(user_id: int, club: schemas.ClubCreate, db: Session) -> schemas.Club:
     club_dict = club.dict()
@@ -54,11 +92,6 @@ def read_club(user_id: int, club_id: int, db: Session) -> schemas.Club:
     )
 
 
-# Club READ without user query
-def read_clubs_for_joining(club_id: int, db: Session) -> schemas.Club:
-    return db.query(models.Club).filter(models.Club.id == club_id).first()
-
-
 # Club READ
 def read_clubs(user_id: int, is_active: bool, db: Session,) -> schemas.Clubs:
     query_results = db.query(models.Club).filter(
@@ -70,8 +103,15 @@ def read_clubs(user_id: int, is_active: bool, db: Session,) -> schemas.Clubs:
 
 
 # Club UPDATE
-def update_club(club_id: int, club: schemas.ClubUpdate, db: Session) -> schemas.Club:
-    db_club = db.query(models.Club).filter(models.Club.id == club_id).first()
+def update_club(
+    user_id: int, club_id: int, club: schemas.ClubUpdate, db: Session
+) -> schemas.Club:
+    db_club = (
+        db.query(models.Club)
+        .filter(models.Club.id == club_id)
+        .filter(models.Club.users.any(models.User.id == user_id))
+        .first()
+    )
     remove_nones = {k: v for k, v in club.dict().items() if v is not None}
     db_club.update(remove_nones)
     db.commit()
@@ -90,8 +130,43 @@ def add_user_to_club(club_id: int, user_id: int, db: Session) -> schemas.Club:
 
 
 # Club UPDATE
-def add_book_to_club(club_id: int, book_id: int, db: Session) -> schemas.Club:
+def remove_user_from_club(club_id: int, user_id: int, db: Session) -> schemas.Club:
     db_club = db.query(models.Club).filter(models.Club.id == club_id).first()
+    if db_club.admin_user_id == user_id:
+        raise HTTPException(status_code=400, detail="Admin cannot leave club")
+    db_user = db.query(models.User).filter(models.User.id == user_id).first()
+    db_club.users.remove(db_user)
+    db.commit()
+    db.refresh(db_club)
+    return db_club
+
+
+# Club UPDATE
+def add_book_to_club(
+    user_id: int, club_id: int, book_id: int, db: Session
+) -> schemas.Club:
+    db_club = (
+        db.query(models.Club)
+        .filter(models.Club.id == club_id)
+        .filter(models.Club.users.any(models.User.id == user_id))
+        .first()
+    )
+    db_book = db.query(models.Book).filter(models.Book.id == book_id).first()
+    db_club.books.append(db_book)
+    db.commit()
+    db.refresh(db_club)
+    return db_club
+
+
+# Club UPDATE
+def remove_book_from_club(
+    user_id: int, club_id: int, book_id: int, db: Session
+) -> schemas.Club:
+    db_club = db.query(models.Club).filter(models.Club.id == club_id).first()
+    if db_club.admin_user_id != user_id:
+        raise HTTPException(
+            status_code=400, detail="Non-admin cannot remove book from club"
+        )
     db_book = db.query(models.Book).filter(models.Book.id == book_id).first()
     db_club.books.append(db_book)
     db.commit()
@@ -100,50 +175,13 @@ def add_book_to_club(club_id: int, book_id: int, db: Session) -> schemas.Club:
 
 
 # Club DELETE
-def delete_club(club_id: int, db: Session) -> schemas.Club:
+def delete_club(user_id: int, club_id: int, db: Session) -> schemas.Club:
     db_club = db.query(models.Club).filter(models.Club.id == club_id).first()
+    if db_club.admin_user_id != user_id:
+        raise HTTPException(status_code=403, detail="Non-admin cannot delete club")
     db.delete(db_club)
     db.commit()
     return db_club
-
-
-# Book CREATE
-def create_book(book: schemas.BookCreate, db: Session) -> schemas.Book:
-    db_book = models.Book(**book.dict())
-    db.add(db_book)
-    db.commit()
-    db.refresh(db_book)
-    return db_book
-
-
-# Book READ
-def read_book(user_id: int, book_id: int, db: Session) -> schemas.Book:
-    db_book = (
-        db.query(models.Book, models.Club)
-        .filter(models.Book.id == book_id)
-        .filter(models.Club.users.any(models.User.id == user_id))
-        .first()
-    )
-    if db_book:
-        return db_book.Book
-
-
-# Book UPDATE
-def update_book(book_id: int, book: schemas.BookUpdate, db: Session) -> schemas.Book:
-    db_book = db.query(models.Book).filter(models.Book.id == book_id).first()
-    remove_nones = {k: v for k, v in book.dict().items() if v is not None}
-    db_book.update(remove_nones)
-    db.commit()
-    db.refresh(db_book)
-    return db_book
-
-
-# Book DELETE
-def delete_book(book_id: int, db: Session) -> schemas.Book:
-    db_book = db.query(models.Book).filter(models.Book.id == book_id).first()
-    db.delete(db_book)
-    db.commit()
-    return db_book
 
 
 # Note CREATE
@@ -164,44 +202,26 @@ def read_note(user_id: int, note_id: int, db: Session) -> schemas.Note:
     and if the note is private, ensure that the user requesting the note
     wrote the note.
     """
-    db_note = (
-        db.query(models.Note, models.Club)
-        .filter(models.Note.id == note_id)
-        .filter(models.Club.users.any(models.User.id == user_id))
-        .first()
-    )
-    if db_note:
-        return db_note.Note
-
-
-# Note READ
-def read_team_notes(
-    user_id: int, club_id: int, archived: bool, db: Session
-) -> schemas.Notes:
-    query_results = (
-        db.query(models.Note, models.Club)
-        .filter(models.Note.club_id == club_id)
-        .filter(models.Club.users.any(models.User.id == user_id))
-        .filter(models.Note.private == False)  # only return public notes
-    )
-    if not archived:
-        query_results = query_results.filter(models.Note.archived == False)
-    results = [qr.Note for qr in query_results.all()]
-    return schemas.Notes(notes=results)
+    db_note = db.query(models.Note).filter(models.Note.id == note_id).first()
+    if db_note.user_id != user_id and db_note.private:
+        raise HTTPException(403, detail="Cannot read private note")
+    return db_note
 
 
 # Note READ
 def read_personal_notes(
-    user_id: int, club_id: int, private: bool, archived: bool, db: Session,
+    user_id: int,
+    book_id: int,
+    include_private: bool,
+    include_archived: bool,
+    db: Session,
 ) -> schemas.Notes:
-    query_results = (
-        db.query(models.Note)
-        .filter(models.Note.user_id == user_id)
-        .filter(models.Note.club_id == club_id)
-    )
-    if not private:
+    query_results = db.query(models.Note).filter(models.Note.user_id == user_id)
+    if book_id:
+        query_results = query_results.filter(models.Note.book_id == book_id)
+    if not include_private:
         query_results = query_results.filter(models.Note.private == False)
-    if not archived:
+    if not include_archived:
         query_results = query_results.filter(models.Note.archived == False)
     return schemas.Notes(notes=query_results.all())
 
@@ -216,6 +236,8 @@ def update_note(
         .filter(models.Note.user_id == user_id)
         .first()
     )
+    if not db_note:
+        raise HTTPException(status_code=400, detail="Note not found")
     remove_nones = {k: v for k, v in note.dict().items() if v is not None}
     db_note.update(remove_nones)
     db.commit()
@@ -225,7 +247,7 @@ def update_note(
 
 # Note UPDATE
 def add_tags_to_note(
-    user_id: int, note_id: int, tag_ids: list, db: Session
+    user_id: int, note_id: int, tags: schemas.NoteAddTagsAndClubTags, db: Session
 ) -> schemas.Note:
     db_note = (
         db.query(models.Note)
@@ -233,9 +255,14 @@ def add_tags_to_note(
         .filter(models.Note.user_id == user_id)
         .first()
     )
-    for tag_id in tag_ids:
+    for tag_id in tags.tags:
         db_tag = db.query(models.Tag).filter(models.Tag.id == tag_id).first()
         db_note.tags.append(db_tag)
+    for club_tag_id in tags.club_tags:
+        db_club_tag = (
+            db.query(models.ClubTag).filter(models.ClubTag.id == club_tag_id).first()
+        )
+        db_note.club_tags.append(db_club_tag)
     db.add(db_note)
     db.commit()
     db.refresh(db_note)
@@ -250,6 +277,8 @@ def delete_note(user_id: int, note_id: int, db: Session) -> schemas.Note:
         .filter(models.Note.id == note_id)
         .first()
     )
+    if not db_note:
+        raise HTTPException(status_code=400, detail="Note not found")
     db.delete(db_note)
     db.commit()
     return db_note
@@ -265,38 +294,40 @@ def create_tag(tag: schemas.TagCreate, db: Session) -> schemas.Tag:
 
 
 # Tag READ
-def read_tag(user_id: int, tag_id: int, db: Session) -> schemas.Tag:
-    db_tag = (
-        db.query(models.Tag, models.User)
-        .filter(models.Tag.id == tag_id)
-        .filter(models.Club.users.any(models.User.id == user_id))
-        .first()
-    )
-    if db_tag:
-        return db_tag.Tag
-
-
-# Tag READ
-def read_duplicate_tag(club_id: int, name: str, db: Session) -> schemas.Tag:
-    return (
+def read_tag(user_id: int, tag_id: int, book_id: int, db: Session) -> schemas.Tag:
+    db_query = (
         db.query(models.Tag)
-        .filter(models.Tag.club_id == club_id)
-        .filter(models.Tag.name == name)
+        .filter(models.Tag.id == tag_id)
+        .filter(models.Tag.user_id == user_id)
         .first()
     )
+    if book_id:
+        db_query.filter(models.Tag.books.any(models.Book.id == book_id))
+    if not db_query:
+        raise HTTPException(status_code=400, detail="Tag not found")
+    return db_query.first()
 
 
 # Tag READ
-def read_tags(club_id: int, archived: bool, db: Session) -> schemas.Tags:
-    query_results = db.query(models.Tag).filter(models.Tag.club_id == club_id)
+def read_tags(user_id: int, archived: bool, db: Session) -> schemas.Tags:
+    query_results = db.query(models.Tag).filter(models.Tag.user_id == user_id)
     if not archived:
         query_results = query_results.filter(models.Tag.archived == False)
     return schemas.Tags(tags=query_results.all())
 
 
 # Tag UPDATE
-def update_tag(tag_id: int, tag: schemas.TagUpdate, db: Session) -> schemas.Tag:
-    db_tag = db.query(models.Tag).filter(models.Tag.id == tag_id).first()
+def update_tag(
+    user_id: int, tag_id: int, tag: schemas.TagUpdate, db: Session
+) -> schemas.Tag:
+    db_tag = (
+        db.query(models.Tag)
+        .filter(models.Tag.id == tag_id)
+        .filter(models.Tag.user_id == user_id)
+        .first()
+    )
+    if not db_tag:
+        raise HTTPException(status_code=400, detail="Tag not found")
     remove_nones = {k: v for k, v in tag.dict().items() if v is not None}
     db_tag.update(remove_nones)
     db.commit()
@@ -305,8 +336,85 @@ def update_tag(tag_id: int, tag: schemas.TagUpdate, db: Session) -> schemas.Tag:
 
 
 # Tag DELETE
-def delete_tag(tag_id: int, db: Session) -> schemas.Tag:
-    db_tag = db.query(models.Tag).filter(models.Tag.id == tag_id).first()
+def delete_tag(user_id: int, tag_id: int, db: Session) -> schemas.Tag:
+    db_tag = (
+        db.query(models.Tag)
+        .filter(models.Tag.id == tag_id)
+        .filter(models.Tag.user_id == user_id)
+        .first()
+    )
+    if not db_tag:
+        raise HTTPException(status_code=400, detail="Tag not found")
     db.delete(db_tag)
     db.commit()
     return db_tag
+
+
+# ClubTag CREATE
+def create_club_tag(club_tag: schemas.TagCreate, db: Session) -> schemas.Tag:
+    db_club_tag = models.Tag(**club_tag.dict())
+    db.add(db_club_tag)
+    db.commit()
+    db.refresh(db_club_tag)
+    return db_club_tag
+
+
+# ClubTag READ
+def read_club_tag(user_id: int, club_tag_id: int, db: Session) -> schemas.Tag:
+    db_query = (
+        db.query(models.ClubTag)
+        .filter(models.ClubTag.id == club_tag_id)
+        .filter(models.ClubTag.user_id == user_id)
+        .first()
+    )
+    if not db_query:
+        raise HTTPException(status_code=400, detail="Tag not found")
+    return db_query
+
+
+# ClubTag READ
+def read_club_tags(
+    user_id: int, book_id: int, archived: bool, db: Session
+) -> schemas.ClubTags:
+    query_results = (
+        db.query(models.ClubTag)
+        .filter(models.ClubTag.book_id == book_id)
+        .filter(models.Club.users.any(models.User.id == user_id))
+    )
+    if not archived:
+        query_results = query_results.filter(models.Tag.archived == False)
+    return schemas.Tags(tags=query_results.all())
+
+
+# ClubTag UPDATE
+def update_club_tag(
+    user_id: int, club_tag_id: int, club_tag: schemas.ClubTagUpdate, db: Session
+) -> schemas.ClubTag:
+    db_club_tag = (
+        db.query(models.ClubTag)
+        .filter(models.ClubTag.id == club_tag_id)
+        .filter(models.ClubTag.club.admin_user_id == user_id)
+        .first()
+    )
+    if not db_club_tag:
+        raise HTTPException(status_code=400, detail="ClubTag not found")
+    remove_nones = {k: v for k, v in club_tag.dict().items() if v is not None}
+    db_club_tag.update(remove_nones)
+    db.commit()
+    db.refresh(db_club_tag)
+    return db_club_tag
+
+
+# ClubTag DELETE
+def delete_club_tag(user_id: int, club_tag_id: int, db: Session) -> schemas.Tag:
+    db_club_tag = (
+        db.query(models.ClubTag)
+        .filter(models.ClubTag.id == club_tag_id)
+        .filter(models.ClubTag.club.admin_user_id == user_id)
+        .first()
+    )
+    if not db_club_tag:
+        raise HTTPException(status_code=400, detail="ClubTag not found")
+    db.delete(db_club_tag)
+    db.commit()
+    return db_club_tag
